feat(image upload): move path helpers
