+++ image/src/upload/upload.service.ts
@@ -12,6 +12,10 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { createHash } from 'crypto';
 import * as sharp from 'sharp';
+import {
+  pathBuilder,
+  validateOrCreateDirectory,
+} from '../shared/helpers/file-path.helpers';
 
 // Code assumes that either UNIX or Windows paths are valid. Any other OS or path format is not supported
 
@@ -21,16 +25,16 @@ export class UploadService {
   private cachePath: string;
   private readonly logger = new Logger(UploadService.name);
   constructor(private readonly configService: ConfigService) {
-    this.storePath = this.pathBuilder(
+    this.storePath = pathBuilder(
       this.configService.get<string>('IMAGE_STORE_BASE_PATH'),
       'disk',
     );
-    this.validateOrCreateDirectory(this.storePath);
-    this.cachePath = this.pathBuilder(
+    validateOrCreateDirectory(this.storePath);
+    this.cachePath = pathBuilder(
       this.configService.get<string>('IMAGE_STORE_INCOMING_CACHE_PATH'),
       'cache',
     );
-    this.validateOrCreateDirectory(this.cachePath);
+    validateOrCreateDirectory(this.cachePath);
   }
   /**
    * @description Handle uploaded image including duplicate check, hashing and saving it
@@ -80,45 +84,6 @@ export class UploadService {
     const fileBuffer = await fsProm.readFile(filePath);
     const hashSum = createHash('sha256').update(fileBuffer);
     return hashSum.digest('hex');
-  }
-
-  /**
-   * @description Take in path and make sure that it´s a global path and not undefined
-   */
-  private pathBuilder(initialPath: string, fallbackFolder: string) {
-    // Set OS specific variables for path handling
-    const absolutePathIndicator = this.getOSAbsolutePathIndicator();
-    // No path? Use local as fallback
-    if (!initialPath) {
-      this.logger.warn('No path passed, so defaulting to fallback');
-      return path.resolve(fallbackFolder);
-    }
-    // Is absolute path
-    if (initialPath.match(absolutePathIndicator)) {
-      return initialPath;
-    }
-    return path.resolve(initialPath);
-  }
-  /**
-   * @description Get Regex to detect absolute paths for the current OS
-   */
-  private getOSAbsolutePathIndicator() {
-    return process.platform.startsWith('win')
-      ? new RegExp('^[A-Z]:.*')
-      : new RegExp('^/.*');
-  }
-  /**
-   * @description Validate a given path to be a directory. If the directory does not yet exist => create it
-   */
-  private validateOrCreateDirectory(path) {
-    if (!fs.existsSync(path)) {
-      fs.mkdirSync(path, { recursive: true });
-    }
-    if (!fs.lstatSync(path).isDirectory()) {
-      throw new Error(
-        `${path} does not point to directory and therefore is not allowed in this context`,
-      );
-    }
   }
 
   /**
